+++ config/hard/Gerrymandering/code/CGCode.cpp
@@ -1,0 +1,24 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    int w;
+    int h;
+    cin >> w >> h; cin.ignore();
+    for (int i = 0; i < h; i++) {
+        for (int j = 0; j < w; j++) {
+            int voters;
+            cin >> voters; cin.ignore();
+        }
+    }
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "answer" << endl;
+}