+++ config/hard/Gerrymandering/code/CGCode.java
@@ -1,0 +1,22 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        int w = in.nextInt();
+        int h = in.nextInt();
+        for (int i = 0; i < h; i++) {
+            for (int j = 0; j < w; j++) {
+                int voters = in.nextInt();
+            }
+        }
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("answer");
+    }
+}