+++ config/hard/Gerrymandering/code/CGCode.ts
@@ -1,0 +1,14 @@
+var inputs: string[] = readline().split(' ');
+const w: number = parseInt(inputs[0]);
+const h: number = parseInt(inputs[1]);
+for (let i = 0; i < h; i++) {
+    var inputs: string[] = readline().split(' ');
+    for (let j = 0; j < w; j++) {
+        const voters: number = parseInt(inputs[j]);
+    }
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');