+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Wave Function Collapse (sans Shannon)".
 - configuration for "Grid climbing".
 - configuration for "Kolakoski Sequence".
+- configuration for "Gerrymandering".
 
 ## [1.15.0] - 2025-05-31
 ### Added